// src/types/restaurantLogin.type.ts
export interface RestaurantLoginResponse{
    id: number;
    email: string;
    role: string;
    restaurantName: string;
    isSubscribed: string;
    isEmailVerified: string;
    restaurantCode: string;
    profilePicture: string;
    bannerImage: string;
    logo: string;
    subscriptionType: string;
    subscriptionEndDate: string;
    isBlocked: string;
    createdAt: string
    address: string;
    country: string;
    phoneNumber: string;

}